+++ RELEASE-NOTES.md
@@ -5,6 +5,8 @@
 - Fixed bug in CSV template that marked all items as flagged.
 - Fixed a ranking bug that caused items to appear in the wrong order in Context mode or after flattening.
 - Improved Markdown formatting.
+- Fixed due/done confusion in Taskpaper template.
+- Fixed HTLM contexts indenting.
 
 **[1.0.11 (2014-12-10)](https://github.com/psidnell/ofexport2/archive/ofexport-v2-1.0.11.zip)**